Visualizing a random walk in python using Random and Matplotlib.
This is a random walk program written in Jupyter Notebook using Python. The implementation was done using 2 Python modules:- 
    1. Random Module (used for generating random numbers)
    2. Matplotlib.pyplot as plt (used for creating visualizations)

Variables and parameters used : 
   1. num_points - Specifies the size of the random walk (how many points are there in random walk)
   2. x_values []- A list to store the x co-ordinates for  random walk.
   3. y_values [] - A list to store the y co-ordinates for random walk.

So, we start by creating a Class RandomWalk(). We can create different objects of this class with different sizes of random walk and visualize them individually. 

Now, the first function inside the class is a constructor and is used to initialize the parameters of the class. The "num_points" parameter is the size of the random walk, here the default value is set of 5000 points. We also set our staring point as (0,0). The start and the endpoint are shown with "Green" and "Red" color points. 

Our Second function, "fill_walk()", is responsible for generating the pattern of the random walk. We start the "While" loop for the value of X until it's greater than "num_points". We choose a direction (Forward or Backward) by selecting between -1 & 1 randomly, and we select the distance for each direction. To calculate the steps in each iteration multiply direction and distance. Do the previous step to generate both x and y steps. After generating the step calculate the next step by adding the calculated steps to the last/latest x and y points. For example, if my last/latest point was (1, 2) and step is (-2, 4) then my next step will be (1+(-2), 2 + 4)= (-1, 6). If step comes out to be 0 then we can use "continue" to skip the rest of the program and move to the next iteration. Next, append the newly calculated values of the next step to the list x_values and y_values.

Lastly, create an object of the class and plot it using plt.scatter (to plot a scatter plot).

Thanks